+++ dcap/quoteverify/src/lib.rs
@@ -6,7 +6,7 @@
 mod quote_enclave;
 mod verify;
 
-use mc_sgx_dcap_types::Quote3Error;
+use mc_sgx_dcap_types::SgxError;
 pub use quote_enclave::{LoadPolicyInitializer, PathInitializer};
 pub use verify::supplemental_data_size;
 
@@ -17,7 +17,7 @@ pub enum Error {
     /// Paths have already been initialized
     PathsInitialized,
     /// Error from SGX quoting library function: {0}
-    Sgx(Quote3Error),
+    Sgx(SgxError),
     /// Failed to convert a path to a string.  Path {0}
     PathStringConversion(String),
     /// Path does not exist
@@ -28,8 +28,8 @@ pub enum Error {
     LoadPolicyInitialized,
 }
 
-impl From<Quote3Error> for Error {
-    fn from(src: Quote3Error) -> Self {
+impl From<SgxError> for Error {
+    fn from(src: SgxError) -> Self {
         Self::Sgx(src)
     }
 }